#128 M3: salvage skill_render (decorations) so the driver is skill_runner-free
Move clear_decorations/attach_diagnostics/highlight_edits out of skill_runner
into skill_render.lua — the single source the M3 skill_invoke driver renders
through, so it doesn't depend on the to-be-deleted skill_runner (ARCH-DRY).
skill_runner delegates (thin pass-throughs, like compute_edits) — v1 path +
spec stay green (9/9). 3 skill_render assertions.
